+++ src/constants/config.ts
@@ -1,4 +1,4 @@
-export const ENVS = {
+export const ENVS: Record<string, string> = {
   WALLET_CONNECT_PROJECT_ID: import.meta.env.VITE_WALLET_CONNECT_PROJECT_ID,
   CHAIN_ENV: import.meta.env.VITE_CHAIN_ENV,
   TG_APP_LINK: import.meta.env.VITE_TMA_LINK, //t.me/ethsignddev_bot/minidev t.me/ChainDevBot/chainapp
