+++ src/main.tsx
@@ -1,35 +1,26 @@
+import { Router } from '@/Router.tsx';
+import { ENVS } from '@/constants/config.ts';
+import { TonProvider } from '@/core/providers/ton';
+import { ThemeProvider, Toaster } from '@ethsign/ui';
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
-import { Router } from '@/Router.tsx';
-import { ThemeProvider, Toaster } from '@ethsign/ui';
-import WebApp from '@twa-dev/sdk';
-import { ENVS } from '@/constants/config.ts';
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { LotteryInfoProvider } from './providers/LotteryInfoProvider';
-import { TonProvider } from '@/core/providers/ton';
 
-WebApp.ready();
-
-WebApp.setBackgroundColor('#fff');
-WebApp.setHeaderColor('#fff');
-
-const manifestUrl = 'https://app.ethsign.xyz/manifest.json';
+// eslint-disable-next-line react-refresh/only-export-components
+const TonConfig = {
+  manifestUrl: 'https://app.ethsign.xyz/manifest.json',
+  actionsConfiguration: {
+    twaReturnUrl: ENVS.TG_APP_LINK as `${string}://${string}`
+  }
+};
 
 const queryClient = new QueryClient();
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
-    <TonProvider
-      config={
-        {
-          manifestUrl: manifestUrl,
-          actionsConfiguration: {
-            twaReturnUrl: ENVS.TG_APP_LINK as any
-          }
-        } as any
-      }
-    >
+    <TonProvider config={TonConfig}>
       <QueryClientProvider client={queryClient}>
         <ThemeProvider defaultTheme={'light'} storageKey={'theme'}>
           <LotteryInfoProvider>
