+++ src/utils/index.ts
@@ -1,7 +1,5 @@
-export function encodeTelegramStartParam(param: unknown) {
-  return encodeURIComponent(window.btoa(JSON.stringify(param)));
-}
-
-export function decodeTelegramStartParam(encodedParam: string) {
-  return JSON.parse(window.atob(decodeURIComponent(encodedParam)));
-}
+export * from './common';
+export * from './lottery';
+export * from './api-client';
+export * from './tailwind';
+export * from './telegram';
